Preparatory refactoring for train and predict CLI
